Refactoring: Moving tests to the correct package europa, not europe
