+++ rest/src/test/java/eu/europa/ec/fisheries/uvms/spatial/rest/ConfigServiceMock.java
@@ -8,7 +8,7 @@ the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the impl
 FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details. You should have received a
 copy of the GNU General Public License along with the IFDM Suite. If not, see <http://www.gnu.org/licenses/>.
  */
-package eu.europe.ec.fisheries.uvms.spatial.rest;
+package eu.europa.ec.fisheries.uvms.spatial.rest;
 
 import java.util.Arrays;
 import javax.ejb.ActivationConfigProperty;
